+++ little_brother/test/test_suite.py
@@ -26,7 +26,6 @@ from little_brother.test import test_client_device_handler
 from little_brother.test import test_client_process_handler
 from little_brother.test import test_german_vacation_context_rule_handler
 from little_brother.test import test_language
-from little_brother.test import test_master_connector
 from little_brother.test import test_process_handler
 from little_brother.test import test_process_handler_manager
 from little_brother.test import test_process_info
@@ -108,10 +107,6 @@ def add_test_cases(p_test_suite, p_config_filename=None):
     base_test.add_tests_in_test_unit(
         p_test_suite=p_test_suite,
         p_test_unit_class=test_user_status.TestUserStatus, p_config_filename=p_config_filename)
-
-    base_test.add_tests_in_test_unit(
-        p_test_suite=p_test_suite,
-        p_test_unit_class=test_master_connector.TestMasterConnector, p_config_filename=p_config_filename)
 
     base_test.add_tests_in_test_unit(
         p_test_suite=p_test_suite,